Create README.md
Excellent app! Actually, in my opinion, the best one! It has a really great salat reminder, it helps you strengthen your Imaan and every day there's a Dua' for you! It has a lot of different Dua's in it, for every situation and it contains the Holy Qur'an in Arabic and translated! You can also read the Qur'an without knowing Arabic so it's the best for me! Masha Allah this app is amazing!